+++ app/templates/api/blueprints/add.js
@@ -1,0 +1,1 @@
+module.exports = require('sails/lib/hooks/blueprints/actions/add');