+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Edge version
 
 - Improvement: Move to Sails v0.11;
+- Improvement: Move to Unstable stability level;
 - Improvement: Move `jwt` config to `JwtCipher` service;
 - Improvement: Add more custom responses `Bad Request`, `Created`, `Forbidden`, `Not Found`, `OK`, `Server Error`, `Unauthorized`;
 - Improvement: You can override root object in response;